base: Make display_list_mutex recursive.
This is kind of an ugly hack, but it allows calling eglInitialize on an
internal eglDisplay during eglGetPlatformDisplay.

The problem is that when we call eglInitialize for an internal display,
the driver will call the queryString export in every vendor library with
EGL_EXT_PLATFORM_DISPLAY_EXTENSIONS to allow vendors to inject
extensions into other EGLDisplays.

eplQueryStringExport tries to call eplDisplayAcquire to check if the
EGLDisplay is one that the platform library owns. That in turn has to
lock display_list_mutex, which will already be locked by
eplGetPlatformDisplayExport.