base: Add helper functions for reference-counted structs.
Define an EplRefCount struct, which is intended to be embedded in a
larger structure.

The functions eplRefCountRef and eplRefCountUnref can be used to
increment and decrement the reference count.

There's also two macros, EPL_REFCOUNT_DECLARE_TYPE_FUNCS and
EPL_REFCOUNT_DEFINE_TYPE_FUNCS, which expand out to type-specific
functions to handle cleaning up a larger struct.