Rework the implementation functions to use a function table.
Replace all of the eplImpl* functions with function pointers in a new
EplImplFuncs struct.

This serves two purposes. First, it allows some of the functions to be
optional.

Second, it means that you could have more than one platform
implementation in the same library, and loadEGLExternalPlatform could
select one at runtime.