base: Add hooks for eglWaitGL and eglWaitNative.
These are a bit awkward, since they don't take an EGLDisplay parameter,
so we'd need to call eglGetPlatformDisplay. But, the pointer to
eglGetPlatformDisplay is in the EplPlatformData struct, which we would
normally look up from an EGLDisplay handle.

So, to deal with that, add a new eplGetCurrentDisplay function, which
iterates over all of the EplPlatformData structs and calls
eglGetCurrentDisplay through each of them until it finds a valid
EGLDisplay.

In practice, there will only ever be one EplPlatformData struct in the
list. But even if the platform library *did* get loaded from multiple
drivers, that would still work because only one of those drivers would
have a current display.

Added new WaitGL and WaitNative pointers to EplImplFuncs. Both are
optional, and if they're NULL, then the base code won't offer the
corresponding hook function to the driver.