X11: Add an XESetCloseDisplay callback.
For Xlib displays, call XESetCloseDisplay to keep track of when the
native display gets closed.

Add a global linked list of structs to keep track of the callbacks.
Since the same Xlib Display can be shared by more than one EGLDisplay,
they'll share the same callback as well.

There's also a reference count for the callback list, in case
loadEGLExternalPlatform gets called more than once.

For eglGetPlatformDisplay, eplImplIsSameDisplay will ignore any existing
EGLDisplays with a closed native display. That way, if the application
calls XCloseDisplay and then XOpenDisplay and happens to get a Display
with the same address, the platform library will still treat it as
different and it will return a new EGLDisplay for it.

In eglInitialize, it will fail with an EGL_BAD_ACCESS error if the
native display is closed.

Those are the only places we check right now. If the application calls
XCloseDisplay before it calls eglCloseDisplay, then things will likely
crash, but that's still arguably an application bug.

Also, this only works with EGL_PLATFORM_X11_KHR, not with
EGL_PLATFORM_XCB_EXT, because libxcb doesn't have any equivalent to
XESetCloseDisplay.