X11: Check support for PRIME and explicit sync
Add the necessary per-display bookkeeping for PRIME. Check for the
necessary format and modifier support, and determine a device to use for
the client, if that's different than what the server is using.

Also add support for EGL_EXT_explicit_device.

Check for the necessary kernel and driver support and protocol versions
for using explicit or implicit sync.