X11: Add support for windows.
Add the necessary functions for rendering to windows.

The platform library should be able to correctly handle both front- and
double-buffered rendering, and should be able to handle both direct and
PRIME for presentation.

For synchronizaiton, it will use explicit sync if that's available.
Otherwise, if it's rendering through PRIME, it'll try to use implicit
sync by attaching fences to the shared dma-buf.

If all else fails, it'll fall back to calling glFinish before
presentation. Corruption is still possible in that case, because we
don't have server -> client sync, but the platform library will try to
use the oldest buffer so that it has the best chance of avoiding that.