X11: Start adding support for PRIME with pixmaps.
If the server is using pitch linear for a format modifier, allocate a
local rendering buffer for a pixmap, and attach the server's buffer to
the EGL_PLATFORM_SURFACE_BLIT_TARGET_NVX attachment point.

This should allow rendering to pixmaps even on a PRIME configuration, as
long as the application calls glFlush, glFinish, or eglWaitGL to ensure
that the blit actually happens.

Also, support may still be limited, because it depends on XWayland
using DRM_FORMAT_MOD_LINEAR when it allocates a pixmap. During local
testing, an XCreatePixmap request ended up with a multi-plane Intel
format.