X11: Add a damage callback for pixmaps
Add a damage callback for pixmap surfaces.

If implicit sync is supported, then the damage callback will attach the
syncfd to the dma-buf. Otherwise, it'll do a CPU wait, which would
produce a functionally equivalent result (albeit at a performance
penalty).

Strictly speaking, EGL doesn't make any synchronization guarantees about
pixmap rendering, so the only way an application can actually ensure
proper synchronization would be to use glFinish or eglWaitGL anyway, or
to use an EGLSync and wait on it in the client.