X11: Support PRIME with non-linear pixmaps
If the server doesn't use a linear buffer for a pixmap, then allocate a
linear intermediate pixmap in the client, and use that for the blit
target.

In the damage callback, send a CopyArea request to copy from the
intermediate pixmap to the application's pixmap.

In FilterNativePixmap, don't restrict pixmaps to linear buffers. As long
as we can create a linear buffer, we can use a CopyArea request to copy
it to whatever the server uses.