X11: Add a runtime version check for libgbm
In eplX11LoadEGLExternalPlatformCommon, check to make sure that we've
got a recent enough version of libgbm.

To do that, check for whether gbm_bo_create_with_modifiers2 exists.
That's the most recently-added GBM function that we need, so if it's
available, then so is everything else.

If that function is missing, then eplX11LoadEGLExternalPlatformCommon
will fail, and the driver will simply skip over the platform library.