X11: Factor out native handle to XID conversion
Add a new eplX11GetNativeXID function, which handles translating a
native handle pointer to an XID.

This function handles the difference between eglCreatePlatform*Surface
and eglCreate*Surface

In the eglCreatePlatform*Surface functions, the native handle is a
pointer to an XID, but the eglCreate*Surface functions, the native
handle is the XID itself. Also, with Xlib, the XID is a (pointer-sized)
unsigned long, but with XCB, it's a uint32_t.

eplX11GetNativeXID handles all of the above.

This also adds a check in eglCreatePlatform*Surface if the native handle
pointer is NULL, so it'll now return an error instead of crashing.