base: Use an rwlock for the surface list.
Use an rwlock to protect the display's surface list.

eglCreate*Surface, eglDestroySurface, and eglTerminate all take the
write lock. All other functions that operate on an EGLSurface take the
read lock.

This will allow most EGL functions that take an EGLSurface to run
concurrently, without worrying about another thread calling
eglDestroySurface or eglTerminate.

That works because, outside of creation and destruction, almost
everything in EplSurface is effectively read-only. The only thing that
might change is the internal EGLSurface handle, and with the new
platform surface interface, that won't need need to change either.

This also changes and simplifies surface destruction. Since
eglSwapBuffers et. al. don't need to worry about another thread
destroying an EplSurface, EplSurface no longer needs to be refcounted.

Now, the EplSurface is always fully destroyed during the
eglDestroySurface hook, and so we can remove EplImplFuncs::FreeSurface.

Also remove eplSurfaceAcquire and eplSurfaceRelease. Instead, there's a
new eplDisplayLockSurfaceList function, which takes the read lock for
the surface list and returns a pointer to the list head, and
eplSurfaceListLookup, which scans the surface list for an external
EGLSurface handle.

As a convenience, eplHookDisplaySurface and eplHookDisplaySurfaceEnd
are wrappers to the functions above to handle looking up an EGLDisplay
and an EGLSurface, with the appropriate locking and unlocking.