Update the X11 code to match the base library changes.
The big change for this is that the base library now uses an rwlock to
protect the surface list and the eglInitialize/eglTerminate state,
instead of a mutex.

Most EGL functions, most importantly eglSwapBuffers, only need to take a
read lock. That allows concurrent calls to eglSwapBuffers, while still
protecting against another thread trying to call eglDestroySurface or
eglTerminate at the same time.

As a result, we don't need a separate "in use" counter for EplDisplay,
EplSurface doesn't need to be refcounted, and we can get rid of all of
the temporary mutex unlock stuff in eglSwapBuffers.

Merged eplX11FreeWindow into eplX11DestroyWindow, since EplSurface isn't
refcounted anymore.

Changed to provide the SwapInterval and QuerySurface functions in
EplImplFuncs instead of providing the entire hook function.

Note that this brings in the base library's stub implementation of
EGL_EXT_buffer_age and EGL_KHR_partial_update, which will just return
zero for EGL_BUFFER_AGE_KHR. Adding a proper implementation of
EGL_BUFFER_AGE_KHR may come later, if one is needed.